+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.9.1
 ## Updates
-* **HTTP Spoofing:** Added `ONLY_WGET_CURL` option in [`monero-bash.conf`](https://github.com/hinto-janaiyo/monero-bash/blob/main/config/monero-bash.conf) to only use randomly selected 2016-2022 versions of `wget/curl` as the [User-Agent.](https://github.com/hinto-janaiyo/monero-bash/blob/main/docs/fake_http_headers/wget_curl) This option and `ONLY_USER_AGENT` will always enabled when sending a non-local RPC call
+* **HTTP Spoofing:** Added `TOR_BROWSER_MIMIC` option that mimics the HTTP headers sent/used by [Tor browser.](https://github.com/hinto-janaiyo/monero-bash/blob/main/docs/fake_http_headers/tor_browser) Tor proxying must be enabled for this option to work
+* **HTTP Spoofing:** Added `ONLY_WGET_CURL` option to only use randomly selected 2016-2022 versions of `wget/curl` as the [User-Agent.](https://github.com/hinto-janaiyo/monero-bash/blob/main/docs/fake_http_headers/wget_curl) This option and `ONLY_USER_AGENT` will always enabled when sending a non-local RPC call
 * **HTTP Spoofing:** [More referers [5] -> [9]](https://github.com/hinto-janaiyo/monero-bash/blob/main/docs/fake_http_headers/referer)
 * **HTTP Spoofing:** [More language permutations [230] -> [4105]](https://github.com/hinto-janaiyo/monero-bash/blob/main/docs/fake_http_headers/language)
 * **Status:** P2Pool stat calculation is slightly faster
@@ -8,7 +9,7 @@
 ## Fixes
 * **Package manager:** The GitHub API returns 1-line JSON (sometimes) which causes parsing errors. This long standing issue was the cause of most update/upgrade failures, it is now fixed
 * **HTTP Spoofing:** Only set once when upgrading multiple packages
-* **Status:** Fixed P2Pool [x/hour, x/day] math
+* **Status:** Fixed P2Pool `x/hour, x/day` math
 * **Watch:** Spacing is auto-adjusted on terminal size changes
 
 ---